backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
grad-cam==3.1.0
fpdf2==2.7.6
librosa==0.10.1
scikit-learn==1.3.2
pydicom==2.4.4
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
google-generativeai==0.3.1
